added seperate role for bots and users in paradise autorole

+++ bot/main.py
@@ -222,15 +222,18 @@ async def on_ready():
 	print("We are online!")
 
 @client.event
-async def on_member_join(member):
+async def on_member_join(member: discord.Member):
 	if member.guild.id == 965285949447753769:
-		role = discord.utils.get(member.guild.roles, name = "Soul Reapers")
-		WelcomeEmbed = discord.Embed(description = f"Hey {member.mention},\nWelcome to Paradise!!\nWe hope you have a great stay in our server",
-									color = lgd.hexConvertor(mn.colorCollection.find({},{"_id":0,"Hex":1})))
-		channel = discord.utils.get(member.guild.text_channels,id = 967822342459904051)
-		ping = await channel.send(member.mention)
-		await ping.delete()
-		await channel.send(embed = WelcomeEmbed)
+		if member.bot:
+			role = discord.utils.get(member.guild.roles, name = "Bots")
+		else:
+			role = discord.utils.get(member.guild.roles, name = "Soul Reapers")
+			WelcomeEmbed = discord.Embed(description = f"Hey {member.mention},\nWelcome to Paradise!!\nWe hope you have a great stay in our server",
+										color = lgd.hexConvertor(mn.colorCollection.find({},{"_id":0,"Hex":1})))
+			channel = discord.utils.get(member.guild.text_channels,id = 967822342459904051)
+			ping = await channel.send(member.mention)
+			await ping.delete()
+			await channel.send(embed = WelcomeEmbed)
 		await member.add_roles(role)
 
 # new, upgraded and personalized help command
